feat: refactor general and alexa
