+++ packages/alexa-types/package.json
@@ -9,6 +9,7 @@
   "dependencies": {
     "@voiceflow/api-sdk": "^1.35.1",
     "@voiceflow/common": "6.5.0",
+    "@voiceflow/base-types": "^1.0.0",
     "@voiceflow/general-types": "^1.54.1",
     "ask-smapi-model": "1.14.0"
   },
