+++ packages/alexa-types/src/constants/index.ts
@@ -1,3 +1,4 @@
 export * from './apl';
+export * from './base';
 export * from './intents';
 export * from './slots';
