+++ packages/alexa-types/src/diagram.ts
@@ -1,5 +1,5 @@
 import { Diagram } from '@voiceflow/api-sdk';
 
-import { AlexaSteps } from './nodes';
+import { AlexaSteps } from './node';
 
 export type AlexaDiagram = Diagram<AlexaSteps>;
