+++ packages/alexa-types/src/index.ts
@@ -1,6 +1,6 @@
 export * from './constants';
 export * from './diagram';
-export * from './nodes';
+export * from './node';
 export * from './program';
 export * from './project';
 export * from './types';
