+++ packages/alexa-types/src/node/constants.ts
@@ -1,5 +1,3 @@
-export { BaseCommand, BaseNode, BasePort, BaseStep } from '@voiceflow/api-sdk';
-
 export enum PermissionType {
   AVS_DISTRIBUTED_AUDIO = 'avs::distributed_audio',
   ALEXA_ASYNC_EVENT_WRITE = 'alexa::async_event:write',
@@ -50,7 +48,6 @@ export enum PermissionType {
 
 export enum NodeType {
   // visuals
-  CARD = 'card',
   DISPLAY = 'display',
 
   // user
@@ -63,12 +60,5 @@ export enum NodeType {
   DEPRECATED = 'deprecated',
 
   // event
-  DIRECTIVE = 'directive',
   EVENT = 'event',
-
-  // request
-  COMMAND = 'command',
-  INTENT = 'intent',
-  INTERACTION = 'interaction',
-  CAPTURE = 'capture',
 }
