+++ packages/alexa-types/src/program.ts
@@ -1,5 +1,5 @@
 import { Program } from '@voiceflow/api-sdk';
 
-import { AlexaCommand, AlexaNodes } from './nodes';
+import { AlexaCommand, AlexaNodes } from './node';
 
 export type AlexaProgram = Program<AlexaNodes, AlexaCommand>;
