+++ packages/alexa-types/src/project/index.ts
@@ -1,4 +1,5 @@
-import { Project } from '@voiceflow/api-sdk';
+import { Member } from '@voiceflow/api-sdk';
+import { Project } from '@voiceflow/base-types';
 
 import { AlexaProjectMemberData } from './member';
 import { AlexaProduct } from './product';
@@ -10,10 +11,13 @@ export interface AlexaProjectData {
   products: Record<string, AlexaProduct>;
 }
 
-export interface AlexaProject extends Project<AlexaProjectData, AlexaProjectMemberData> {
+export interface AlexaProject extends Project.Project {
+  members: Member<AlexaProjectMemberData>[];
   platform: 'alexa';
+  platformData: AlexaProjectData;
 }
 
-export const defaultAlexaProjectData = ({ products = {} }: Partial<AlexaProjectData> = {}): AlexaProjectData => ({
+export const defaultAlexaProjectData = ({ products = {}, ...data }: Partial<AlexaProjectData> = {}): AlexaProjectData => ({
+  ...Project.defaultProjectData(data),
   products,
 });
