+++ packages/alexa-types/src/version/index.ts
@@ -1,8 +1,8 @@
-import { StrictVersionPlatformData, Version } from '@voiceflow/api-sdk';
-import { defaultBaseVersionData, Locale } from '@voiceflow/general-types';
+import { VersionPrototype } from '@voiceflow/api-sdk';
+import { Locale, Version as GeneralVersion } from '@voiceflow/general-types';
 
-import { AlexaCommand } from '@/nodes';
-import { Voice } from '@/types';
+import { Voice } from '@/constants';
+import { AlexaCommand } from '@/node';
 
 import { AlexaVersionPublishing, defaultAlexaVersionPublishing } from './publishing';
 import { AlexaVersionSettings, defaultAlexaVersionSettings } from './settings';
@@ -16,13 +16,16 @@ export enum AlexaStage {
   REVIEW = 'REVIEW',
 }
 
-export interface AlexaVersionData extends StrictVersionPlatformData<AlexaVersionSettings, AlexaVersionPublishing> {
-  status: {
-    stage: AlexaStage;
-  };
+export interface AlexaVersionData extends GeneralVersion.BaseVersionData<Voice> {
+  status: { stage: AlexaStage };
+  settings: AlexaVersionSettings;
+  publishing: AlexaVersionPublishing;
 }
 
-export type AlexaVersion = Version<AlexaVersionData, AlexaCommand, Locale>;
+export interface AlexaVersion extends GeneralVersion.BaseVersion<Voice> {
+  prototype: VersionPrototype<AlexaCommand, Locale>;
+  platformData: AlexaVersionData;
+}
 
 export const defaultAlexaVersionData = ({
   status: { stage = AlexaStage.DEV } = { stage: AlexaStage.DEV },
@@ -30,7 +33,7 @@ export const defaultAlexaVersionData = ({
   publishing,
   ...generalVersionData
 }: Partial<AlexaVersionData>): AlexaVersionData => ({
-  ...defaultBaseVersionData<Voice>(generalVersionData, { defaultPromptVoice: Voice.ALEXA }),
+  ...GeneralVersion.defaultBaseVersionData<Voice>(generalVersionData, { defaultPromptVoice: Voice.ALEXA }),
   status: { stage },
   settings: defaultAlexaVersionSettings(settings),
   publishing: defaultAlexaVersionPublishing(publishing),
