+++ packages/alexa-types/src/version/publishing.ts
@@ -1,4 +1,4 @@
-import { Locale } from '@/types';
+import { Locale } from '@/constants';
 
 export interface AlexaVersionPublishing {
   forExport: boolean;
