+++ packages/base-types/src/version/index.ts
@@ -1,11 +1,13 @@
-import { StrictVersionPlatformData } from '@voiceflow/api-sdk';
+/* eslint-disable @typescript-eslint/no-empty-interface */
+import { StrictVersionPlatformData, Version as BaseVersion } from '@voiceflow/api-sdk';
 
 import { defaultVersionSettings, VersionSettings } from './settings';
 
 export * from './settings';
 
-// eslint-disable-next-line @typescript-eslint/no-empty-interface
 export interface VersionData extends StrictVersionPlatformData<VersionSettings> {}
+
+export interface Version extends BaseVersion<VersionData> {}
 
 export const defaultVersionData = ({ slots = [], intents = [], settings, publishing = {} }: Partial<VersionData>): VersionData => ({
   slots,
