+++ packages/general-types/src/constants/base.ts
