+++ packages/general-types/src/constants/index.ts
@@ -1,4 +1,5 @@
 export * from './azureVoices';
+export * from './base';
 export * from './device';
 export * from './intent';
 export * from './slot';
