+++ packages/general-types/src/constants/intent.ts
@@ -1,4 +1,4 @@
-import { Language } from '@/constants';
+import { Language } from './base';
 
 export enum IntentName {
   NO = 'VF.NO',
