+++ packages/general-types/src/node/capture.ts
@@ -1,11 +1,11 @@
 /* eslint-disable @typescript-eslint/no-empty-interface */
 
-import { Node } from '@voiceflow/base-types';
+import { Button, Node, Request } from '@voiceflow/base-types';
 
 import { NodeReprompt, StepReprompt } from './utils';
 
-export interface StepData<V> extends Node.Capture.StepData, StepReprompt<V> {}
+export interface StepData extends Node.Capture.StepData, Button.StepButton, StepReprompt {}
 
-export interface Step<V> extends Node.Capture.Step<StepData<V>> {}
+export interface Step extends Node.Capture.Step<StepData> {}
 
-export interface Node extends Node.Capture.Node, NodeReprompt {}
+export interface Node extends Node.Capture.Node, Request.NodeButton, NodeReprompt {}
