+++ packages/general-types/src/node/index.ts
@@ -1,7 +1,5 @@
 import { Node } from '@voiceflow/base-types';
 
-import { Voice } from '@/constants';
-
 import * as Capture from './capture';
 import * as Interaction from './interaction';
 import * as Prompt from './prompt';
@@ -15,11 +13,5 @@ export * as Speak from './speak';
 type OverloadedBaseSteps = Node.Capture.Step | Node.Interaction.Step | Node.Prompt.Step | Node.Speak.Step;
 type OverloadedBaseNodes = Node.Capture.Node | Node.Interaction.Node | Node.Speak.Node;
 
-type BaseSteps = Exclude<Node.AnyBaseStep, OverloadedBaseSteps>;
-type BaseNodes = Exclude<Node.AnyBaseNode, OverloadedBaseNodes>;
-
-export type GeneralBaseSteps<V> = BaseSteps | Speak.Step<V> | Prompt.Step<V> | Capture.Step<V> | Interaction.Step<V>;
-export type GeneralBaseNodes = BaseNodes | Speak.Node | Capture.Node | Interaction.Node;
-
-export type GeneralSteps = GeneralBaseSteps<Voice>;
-export type GeneralNodes = BaseNodes;
+export type GeneralSteps = Exclude<Node.AnyBaseStep, OverloadedBaseSteps> | Speak.Step | Prompt.Step | Capture.Step | Interaction.Step;
+export type GeneralNodes = Exclude<Node.AnyBaseNode, OverloadedBaseNodes> | Speak.Node | Capture.Node | Interaction.Node;
