+++ packages/general-types/src/node/interaction.ts
@@ -1,13 +1,13 @@
 /* eslint-disable @typescript-eslint/no-empty-interface */
 
-import { Node } from '@voiceflow/base-types';
+import { Button, Node, Request } from '@voiceflow/base-types';
 
 import { NodeNoMatch, NodeReprompt, StepNoMatch, StepReprompt } from './utils';
 
-export interface StepData<V> extends Node.Interaction.StepData, StepReprompt<V> {
-  else: StepNoMatch<V>;
+export interface StepData extends Node.Interaction.StepData, Button.StepButton, StepReprompt {
+  else: StepNoMatch;
 }
 
-export interface Step<V> extends Node.Interaction.Step<StepData<V>> {}
+export interface Step extends Node.Interaction.Step<StepData> {}
 
-export interface Node<E = Node.Utils.BaseEvent> extends Node.Interaction.Node<E>, NodeNoMatch, NodeReprompt {}
+export interface Node<E = Node.Utils.BaseEvent> extends Node.Interaction.Node<E>, Request.NodeButton, NodeNoMatch, NodeReprompt {}
