+++ packages/general-types/src/node/prompt.ts
@@ -1,11 +1,11 @@
 /* eslint-disable @typescript-eslint/no-empty-interface */
 
-import { Node } from '@voiceflow/base-types';
+import { Button, Node } from '@voiceflow/base-types';
 
 import { StepNoMatch, StepReprompt } from './utils';
 
-export interface StepData<V> extends Node.Prompt.StepData, StepReprompt<V> {
-  noMatches: StepNoMatch<V>;
+export interface StepData extends Node.Prompt.StepData, Button.StepButton, StepReprompt {
+  noMatches: StepNoMatch;
 }
 
-export interface Step<V> extends Node.Prompt.Step<StepData<V>> {}
+export interface Step extends Node.Prompt.Step<StepData> {}
