+++ packages/general-types/src/node/speak.ts
@@ -2,11 +2,12 @@
 
 import { Node } from '@voiceflow/base-types';
 
+import { Voice } from '@/constants';
 import { Prompt } from '@/types';
 
-export interface StepData<V> extends Node.Speak.StepData, Node.Speak.StepDataDialog<Prompt<V>[]> {}
+export interface StepData extends Node.Speak.StepData, Node.Speak.StepDataDialog<Prompt<Voice>[]> {}
 
-export interface Step<V> extends Node.Speak.Step<StepData<V>> {}
+export interface Step extends Node.Speak.Step<StepData> {}
 
 export interface SpeakNode extends Node.Speak.Node, Node.Speak.SpeakNode<string> {}
